+++ NLCentralStandings/NLCentralStandings/NLCentralStandingsApp.swift
@@ -1,0 +1,12 @@
+import SwiftUI
+
+@main
+struct NLCentralStandingsApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+        .defaultSize(width: 420, height: 520)
+        .windowResizability(.contentSize)
+    }
+}